Relaxed dependencies to reduce 'composer global require' conflicts.
Composer global require simply adds a dependency to a regular project stored in ~/.composer
This means that any packages you have installed in this way need to have internally compatible
dependencies. If dependencies on packages are too strict, you risk conflicts. Because of this,
I have relaxed the requirements of this package, as our use-cases for the console and process
packages are pretty simple.